committed first functional pr

+++ modules/data_release/ajax/FileUpload.php
@@ -25,6 +25,7 @@ if (!$user->hasPermission('data_release_upload')) {
 if ($_GET['action'] == 'upload') {
     $fileName    = $_FILES["file"]["name"];
     $version     = !empty($_POST['version']) ? $_POST['version'] : null;
+    $overwrite   = $_GET['overwrite'] ?? false;
     $upload_date = date('Y-m-d');
 
     $settings = $factory->settings();
@@ -40,32 +41,27 @@ if ($_GET['action'] == 'upload') {
         error_log('ERROR: ' . $msg);
         header("HTTP/1.1 500 Internal Server Error");
     } else {
-        $files = $DB->pselect(
-            "SELECT id, file_name FROM data_release",
+        $fileNames = $DB->pselectcol(
+            "SELECT file_name FROM data_release",
             array()
         );
-        $duplicates = [];
-        foreach ($files as $file) {
-            if ($fileName == $file['file_name']) {
-                $duplicates[$file['id']] = $file['file_name'];
-            }
+
+        if (in_array($fileName, $fileNames) && !$overwrite) {
+            header("Location: {$baseURL}/data_release/?duplicate=true");
+            exit;
         }
 
         $target_path = $path . $fileName;
         if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_path)) {
-            if (!empty($duplicates)) {
-                $DB->delete('data_release', ['file_name' => $fileName]);
-                foreach ($duplicates as $id => $name) {
-                    $DB->update(
-                        'data_release',
-                        [
-                         'file_name'   => $fileName,
-                         'version'     => $version,
-                         'upload_date' => $upload_date,
-                        ],
-                        ['id' => $id] 
-                    );
-                }
+            if (in_array($fileName, $fileNames)) {
+                $DB->update(
+                    'data_release',
+                    [
+                        'version'     => $version,
+                        'upload_date' => $upload_date,
+                    ],
+                    ['file_name' => $fileName]
+                );
             } else {
                 // insert the file into the data_release table
                 $DB->insert(
@@ -76,37 +72,38 @@ if ($_GET['action'] == 'upload') {
                         'upload_date' => $upload_date,
                     )
                 );
+                // get the ID of the user who uploaded the file
+                $user_ID = $DB->pselectOne(
+                    "SELECT ID FROM users WHERE userid=:UserID",
+                    array('UserID' => $user->getUsername())
+                );
+                // get the ID of the file inserted in the data_release table
+                $version_where = $version ? "version=:version"
+                    : "version IS :version";
+                $ID            = $DB->pselectOne(
+                    "SELECT 
+                   id 
+                 FROM 
+                   data_release 
+                 WHERE 
+                   file_name=:file_name 
+                   AND $version_where
+                   AND upload_date=:upload_date",
+                    array(
+                        'file_name'   => $fileName,
+                        'version'     => $version,
+                        'upload_date' => $upload_date,
+                    )
+                );
+                // add permission to the user for the uploaded data_release file
+                $DB->insert(
+                    'data_release_permissions',
+                    array(
+                        'userid'          => $user_ID,
+                        'data_release_id' => $ID,
+                    )
+                );
             }
-            // get the ID of the user who uploaded the file
-            $user_ID = $DB->pselectOne(
-                "SELECT ID FROM users WHERE userid=:UserID",
-                array('UserID' => $user->getUsername())
-            );
-            // get the ID of the file inserted in the data_release table
-            $version_where = $version ? "version=:version" : "version IS :version";
-            $ID            = $DB->pselectOne(
-                "SELECT 
-               id 
-             FROM 
-               data_release 
-             WHERE 
-               file_name=:file_name 
-               AND $version_where
-               AND upload_date=:upload_date",
-                array(
-                    'file_name'   => $fileName,
-                    'version'     => $version,
-                    'upload_date' => $upload_date,
-                )
-            );
-            // add permission to the user for the uploaded data_release file
-            $DB->insert(
-                'data_release_permissions',
-                array(
-                    'userid'          => $user_ID,
-                    'data_release_id' => $ID,
-                )
-            );
         }
         header("Location: {$baseURL}/data_release/?uploadSuccess=true");
         header("HTTP/1.1 201 Created");
